List dismiss 완성
