+++ app/src/main/res/layout/activity_list.xml
@@ -27,12 +27,12 @@
     <LinearLayout
         android:orientation="horizontal"
         android:layout_width="match_parent"
-        android:layout_height="470dp">
+        android:layout_height="420dp">
 
         <ListView
             android:id="@+id/mlist"
             android:layout_width="wrap_content"
-            android:layout_height="match_parent">
+            android:layout_height="420dp">
         </ListView>
     </LinearLayout>
