+++ app/src/main/res/layout/listview_list_item.xml
@@ -16,33 +16,39 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:textSize="20dp"
-            android:paddingLeft="30dp" />
+            android:paddingLeft="30dp"
+            android:gravity="center_vertical"/>
         <TextView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:text="+"
             android:textSize="20dp"
-            android:paddingLeft="30dp" />
+            android:paddingLeft="30dp"
+            android:gravity="center_vertical"/>
         <TextView
             android:id="@+id/second_setting"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:textSize="20dp"
-            android:paddingLeft="30dp" />
+            android:paddingLeft="30dp"
+            android:gravity="center_vertical"/>
         <TextView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:text="|"
             android:textSize="20dp"
-            android:paddingLeft="30dp" />
+            android:paddingLeft="30dp"
+            android:gravity="center_vertical"/>
+        <ImageView
+            android:id="@+id/list_icon"
+            android:layout_width="63dp"
+            android:layout_height="45dp"
+            android:src="@mipmap/kakao"
+            android:paddingLeft="30dp"
+            android:gravity="center_vertical"/>
 
 
     </LinearLayout>
 
-    <ImageView
-        android:id="@+id/list_icon"
-        android:layout_width="63dp"
-        android:layout_height="45dp"
-        android:src="@mipmap/kakao"
-        android:paddingLeft="30dp" />
+
 </LinearLayout>
